Using miner in rpc instead of sync
